Migrating section/sectionGroup circular dependencies from tasks to scheduleOpsService
