+++ src/main/java/edu/ucdavis/dss/ipa/api/components/course/CourseViewController.java
@@ -475,7 +475,7 @@ public class CourseViewController {
 					sectionGroup = sectionGroupService.save(sectionGroup);
 
 					// Find or create a section
-					Section section = sectionService.findOrCreateBySectionGroupIdAndSequenceNumber(sectionGroup.getId(), dwSection.getSequenceNumber());
+					Section section = sectionService.findOrCreateBySectionGroupAndSequenceNumber(sectionGroup, dwSection.getSequenceNumber());
 
 					section.setSeats(dwSection.getMaximumEnrollment());
 					section = sectionService.save(section);
@@ -622,7 +622,7 @@ public class CourseViewController {
 				newSectionGroup = sectionGroupService.save(newSectionGroup);
 
 				for (Section historicalSection : historicalSectionGroup.getSections()) {
-					Section newSection = sectionService.findOrCreateBySectionGroupIdAndSequenceNumber(newSectionGroup.getId(), historicalSection.getSequenceNumber());
+					Section newSection = sectionService.findOrCreateBySectionGroupAndSequenceNumber(newSectionGroup, historicalSection.getSequenceNumber());
 					newSection.setSeats(historicalSection.getSeats());
 					newSection = sectionService.save(newSection);
